+++ Source/DebitsController.m
@@ -730,11 +730,11 @@ extern NSString *DebitReadyForUseDataType;        // For dragging an edited tran
     NSManagedObjectContext *context = MOAssistant.sharedAssistant.context;
 
     if (type == DebitReadyForUseDataType) {
-        int res = NSRunAlertPanel(NSLocalizedString(@"AP417", nil),
-                                  NSLocalizedString(@"AP419", nil),
-                                  NSLocalizedString(@"AP2", nil),
-                                  NSLocalizedString(@"AP10", nil),
-                                  nil);
+        NSInteger res = NSRunAlertPanel(NSLocalizedString(@"AP417", nil),
+                                        NSLocalizedString(@"AP419", nil),
+                                        NSLocalizedString(@"AP2", nil),
+                                        NSLocalizedString(@"AP10", nil),
+                                        nil);
         if (res != NSAlertAlternateReturn) {
             return NO;
         }
@@ -763,11 +763,11 @@ extern NSString *DebitReadyForUseDataType;        // For dragging an edited tran
         warningTitle = (entries.count == 1) ? NSLocalizedString(@"AP421", nil) : NSLocalizedString(@"AP422", nil);
     }
     NSString *warningText = (entries.count == 1) ? NSLocalizedString(@"AP419", nil) : NSLocalizedString(@"AP420", nil);
-    int      res = NSRunAlertPanel(NSLocalizedString(warningTitle, nil),
-                                   NSLocalizedString(warningText, nil),
-                                   NSLocalizedString(@"AP2", nil),
-                                   NSLocalizedString(@"AP10", nil),
-                                   nil);
+    NSInteger res = NSRunAlertPanel(NSLocalizedString(warningTitle, nil),
+                                    NSLocalizedString(warningText, nil),
+                                    NSLocalizedString(@"AP2", nil),
+                                    NSLocalizedString(@"AP10", nil),
+                                    nil);
     if (res != NSAlertAlternateReturn) {
         return NO;
     }
@@ -1074,11 +1074,11 @@ extern NSString *DebitReadyForUseDataType;        // For dragging an edited tran
 
     NSManagedObjectContext *context = MOAssistant.sharedAssistant.context;
 
-    int res = NSRunAlertPanel(NSLocalizedString(@"AP417", nil),
-                              NSLocalizedString(@"AP419", nil),
-                              NSLocalizedString(@"AP2", nil),
-                              NSLocalizedString(@"AP10", nil),
-                              nil);
+    NSInteger res = NSRunAlertPanel(NSLocalizedString(@"AP417", nil),
+                                    NSLocalizedString(@"AP419", nil),
+                                    NSLocalizedString(@"AP2", nil),
+                                    NSLocalizedString(@"AP10", nil),
+                                    nil);
     if (res != NSAlertAlternateReturn) {
         return;
     }
@@ -1173,11 +1173,11 @@ extern NSString *DebitReadyForUseDataType;        // For dragging an edited tran
     NSString *warningTitle;
     warningTitle = (selection.count == 1) ? NSLocalizedString(@"AP421", nil) : NSLocalizedString(@"AP422", nil);
     NSString *warningText = (selection.count == 1) ? NSLocalizedString(@"AP419", nil) : NSLocalizedString(@"AP420", nil);
-    int      res = NSRunAlertPanel(NSLocalizedString(warningTitle, nil),
-                                   NSLocalizedString(warningText, nil),
-                                   NSLocalizedString(@"AP2", nil),
-                                   NSLocalizedString(@"AP10", nil),
-                                   nil);
+    NSInteger res = NSRunAlertPanel(NSLocalizedString(warningTitle, nil),
+                                    NSLocalizedString(warningText, nil),
+                                    NSLocalizedString(@"AP2", nil),
+                                    NSLocalizedString(@"AP10", nil),
+                                    nil);
     if (res != NSAlertAlternateReturn) {
         return;
     }
