+++ Source/Objective-Zip/MiniZip/unzip.c
@@ -670,7 +670,7 @@ local int unzGoToNextDisk(unzFile file)
         }
         else
         {
-            s->filestream = ZOPENDISK64(s->z_filefunc, s->filestream_with_CD, number_disk_next, 
+            s->filestream = ZOPENDISK64(s->z_filefunc, s->filestream_with_CD, (unsigned)number_disk_next,
                 ZLIB_FILEFUNC_MODE_READ | ZLIB_FILEFUNC_MODE_EXISTING);
         }
 
@@ -1393,8 +1393,7 @@ extern int ZEXPORT unzReadCurrentFile(unzFile file, voidp buf, unsigned len)
             uInt total_bytes_read = 0;
 
             if (pfile_in_zip_read_info->stream.next_in != NULL)
-                bytes_not_read = pfile_in_zip_read_info->read_buffer + UNZ_BUFSIZE - 
-                    pfile_in_zip_read_info->stream.next_in;
+                bytes_not_read = (unsigned)pfile_in_zip_read_info->read_buffer + UNZ_BUFSIZE - (unsigned)pfile_in_zip_read_info->stream.next_in;
             bytes_to_read -= bytes_not_read;
             if (bytes_not_read > 0)
                 memcpy(pfile_in_zip_read_info->read_buffer, pfile_in_zip_read_info->stream.next_in, bytes_not_read);
@@ -1410,7 +1409,7 @@ extern int ZEXPORT unzReadCurrentFile(unzFile file, voidp buf, unsigned len)
                         ZLIB_FILEFUNC_SEEK_SET) != 0)
                     return UNZ_ERRNO;
 
-                bytes_read = ZREAD64(pfile_in_zip_read_info->z_filefunc, pfile_in_zip_read_info->filestream,
+                bytes_read = (unsigned)ZREAD64(pfile_in_zip_read_info->z_filefunc, pfile_in_zip_read_info->filestream,
                           pfile_in_zip_read_info->read_buffer + bytes_not_read + total_bytes_read,
                           bytes_to_read - total_bytes_read);
